green_mission
[[https://github.com/username/repository/blob/master/img/octocat.png|alt=octocat]]

The Green Mission is a computer game that is built around the concept of bioremediation (the method using specific organisms to absorb and degrade toxic chemical from the environment).The idea of the game came from Plants vs. Zombies, but instead of having Zombies as the antagonis - the game has the mutated monsters that came from human waste as the enermies. This game won “Best Application Award” from CodeDay Corvallis in January, 2014.